Alpha-Update-1
This update includes the school stats manager which is attached to the dropdown list on the UI. This update also includes the ability to purchase class upgrades, the one currently working is the the advanced literature class. It turns on one of the classes in timetable list, Monday AM, needs a way to turn them all on.