Alpha-Update-Nightwork
I created an Integer on Jacks Student Spawner Script for the current number of students, its equal to the in_SpawmAmount. As the divide for the avarage morality and happiness were incorrect when using the array size itself.

I've finished the background audio script now thanks to the correct averages being created, with it changing between each song.

The Random Event Template has been created and spawns in the scene. The hero and villian decks are now added when the year end if requirements are met, function wasn't working before and is now.

The Random Events now pause the game, Jacks pause form the JH_Time_UI script is now attached to Ovi Random Event script, so when an event is activated it pauses the game. Also clicking on the close button on the random event starts the game again, again uses Jacks script.

One problem that arose is that the coroutine for the random event timer doesnt start up again even with a function just for starting it.

I've confirmed that the AC_YearEnd script works, graduation function checks if stats are met and increases super number and so on.